Update answers.sql
the different departments for which a worker has worked must be counted, since it may be the case that a worker works in department 'd005' then in department 'd004' and then returns to work in department 'd005' another time, then the correct answer to the question in how many departments have you worked? would be 2, while count(de.dept_no) would return 3.